Removed unnecessary ".0" instances when printing.
Time remaining will now not show "12.0 hours, 5.0 minutes," but instead skip the trailing zero-valued decimals.

+++ time_remaining_printer.py
@@ -20,20 +20,20 @@ def show_user_time_remaining(seconds):
         print(str(calc_m)+" minute(s), "+str(round(calc_s,2))+" seconds.")
     elif calc < 86400.0:
         calc_h = int(calc // 3600)
-        calc_m = (calc -(calc_h * 3600)) // 60
+        calc_m = int((calc -(calc_h * 3600)) // 60)
         calc_s = calc -(calc_h * 3600) -(calc_m * 60)
         print(str(calc_h)+" hour(s), "+str(calc_m)+" minutes, "+str(round(calc_s,2))+" seconds.")
     elif calc < 604800:
         calc_d = int(calc // 86400)
-        calc_h = (calc -(calc_d * 86400)) // 3600
-        calc_m = (calc -(calc_d * 86400) -(calc_h * 3600)) // 60
+        calc_h = int((calc -(calc_d * 86400)) // 3600)
+        calc_m = int((calc -(calc_d * 86400) -(calc_h * 3600)) // 60)
         calc_s =  calc -(calc_d * 86400) -(calc_h * 3600) -(calc_m * 60)
         print(str(calc_d)+" day(s), "+str(calc_h)+" hours, "+str(calc_m)+" minutes, "+str(round(calc_s,2))+" seconds.")
     elif calc < 2629743.83:
         calc_w = int(calc // 604800)
-        calc_d = (calc -(calc_w * 604800)) // 86400
-        calc_h = (calc -(calc_w * 604800) -(calc_d * 86400)) // 3600
-        calc_m = (calc -(calc_w * 604800) -(calc_d * 86400) -(calc_h * 3600)) // 60
+        calc_d = int((calc -(calc_w * 604800)) // 86400)
+        calc_h = int((calc -(calc_w * 604800) -(calc_d * 86400)) // 3600)
+        calc_m = int((calc -(calc_w * 604800) -(calc_d * 86400) -(calc_h * 3600)) // 60)
         calc_s =  calc -(calc_w * 604800) -(calc_d * 86400) -(calc_h * 3600) -(calc_m * 60)
         print(str(calc_w)+" week(s), "+str(calc_d)+" days, "+str(calc_h)+" hours, "+str(calc_m)+" minutes, "+str(round(calc_s,2))+" seconds.")
     print("##############################\n")
